last update 15/2

+++ src/main/java/com/example/CorporateLearningManagmentSystem/entity/User.java
@@ -11,24 +11,24 @@ import java.util.Set;
 public class User {
  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
- private Long id;
+ private int id;
  private String name;
  private String email;
  private String password;
 
 
-    public User(Long id, String name, String email, String password) {
+    public User(int id, String name, String email, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.password = password;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
